# app/docs/swagger.yaml
basePath: /
definitions:
  dto.AddShortenDto:
    properties:
      custom_url:
        type: string
      exp:
        default: 0
        type: integer
      url:
        type: string
    required:
    - url
    type: object
  dto.AddUserDto:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 5
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  utils.ApiResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
  utils.ValidationError:
    properties:
      field:
        type: string
      namespace:
        type: string
      tag:
        type: string
      type:
        type: string
      value: {}
    type: object
info:
  contact: {}
  description: This is scissor server.
paths:
  /{id}:
    get:
      consumes:
      - application/json
      description: Redirect or Resolve Url.
      parameters:
      - description: Shorten ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/utils.ValidationError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Resolve Url.
      tags:
      - Shorten
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login User Account.
      parameters:
      - description: Login Dto
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/utils.ValidationError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Login User.
      tags:
      - Auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Create User Account.
      parameters:
      - description: Add User Dto
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.AddUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/utils.ValidationError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Create User.
      tags:
      - Auth
  /api/v1/shorten:
    post:
      consumes:
      - application/json
      description: New Shorten Url.
      parameters:
      - description: Add ShortenDto
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.AddShortenDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AddShortenDto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/utils.ValidationError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Shorten Url.
      tags:
      - Shorten
  /api/v1/user/history:
    get:
      consumes:
      - application/json
      description: History of Shorten Url.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AddShortenDto'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/utils.ValidationError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: History of Shorten Url.
      tags:
      - User
  /api/v1/user/history/:id:
    post:
      consumes:
      - application/json
      description: Delete Shorten Url.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/utils.ValidationError'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Delete Shorten Url.
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"